requirements.txt
# Core dependencies
streamlit>=1.28.0
flask>=2.3.0
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0

# Database
cassandra-driver>=3.25.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization
matplotlib>=3.6.0
plotly>=5.15.0

# Optional: For better error handling and logging
urllib3>=1.26.0
certifi>=2022.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0